Modularize handlers
