+++ src/handlers/mod.rs
@@ -1,0 +1,5 @@
+pub mod download;
+pub mod execute;
+pub mod ping;
+pub mod upload;
+pub mod welcome;
