+++ src/handlers/ping.rs
@@ -1,0 +1,6 @@
+use actix_web::HttpResponse;
+
+pub async fn handler() -> HttpResponse {
+    HttpResponse::Ok().body("pong")
+}
+
